+++ spatial-utilities/src/main/java/org/h2gis/utilities/jts_utils/CoordinateSequenceDimensionFilter.java
@@ -22,7 +22,7 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.h2gis.utilities.geometryUtils.filter;
+package org.h2gis.utilities.jts_utils;
 
 import com.vividsolutions.jts.geom.CoordinateSequence;
 import com.vividsolutions.jts.geom.CoordinateSequenceFilter;
